Update audio player (stories)
Rename "isPlaying" prop to "shouldPlay" to better communicate the
intent of functionality in the <PureAudioPlayer/> component. I.e,
the reducer's "isPlaying" state is used to "sync" the behavior of
the "actual" audio player (playing or not playing).
So this state prope drives the actual audio player state, but does
not necessarily specify if the player	is "actually" playing--this
is denoted by the "isPaused".
In other words, the reducer's "isPlaying" can be thought of as the
desired state, the hook's "isPaused" is the actual state.

And update the stories.