Remove theme config
Fixes #36

Currently no "advanced" theming is used, and all variable changes
are applied through CSS vars, which can easily be changed in one
place (src/global.css) without changing the code in all components.

For now this works well enough, and we'll reevaluate this
approach if requirements change, or if it becomes too "painful" to
work with.