Disable e2e testing
There was one test for the failed login flow, but it only worked because
of a bug (that was fixed in #91).

I think it's better to take a step back, and rethink the best way
to e2e test the app (because of the implications of the app's auth flow
using PKCE).

Therefore I'm temporarily disabling e2e in the CI/CD pipeline, and deleting the
"login failed" e2e test. This test will be converted to a component unit test
instead.

For more information see:
https://github.com/upstandfm/app/issues/94