Clean audio player file resources that don't have to be cached
This change prevents storing update recording in memory, when the <NewUpdate />
component unmounts. These are "cheap" resources we don't have to cache (in contrast
to the audio files that have to be downloaded using a signed download URL).

Additionally, we now resest the playing file when <NewUpdate /> mounts, so a user
can immediately start recording their update without being distracted/bothered by
any audio that was playing before.