Rename "update-recordings" module to "recordings"
Rename the components to better communicate what they render. And update the
"export API".

The updates data we fetch is "generic", and we have a specific usecase (view)
which consists of grouping recordings by user (i.e. by workspace member).
This view (<RecordingsByMember/>) consists of another generic component <Recording/>,
which is the recording itself.

With this change, if we would like to (also) display a list of all updates in chronological
order (i.e. a flat list), we can just create a new component and reuse <Recording/>.