Remove "standup-members" module
This module became superfluous because of the new data model changes
that introduce a workspace, and make users (i.e. members) part of said
workspace.
The "workspace-members" module is effectively a replacement for this one.

Additionaly, we also fix the "leaking API" of <Avatar/>, by no longer exporting
"internal" components.